# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TeamHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teams/{teamId}/events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events for a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event for a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}/attendance:
    get:
      operationId: listAttendance
      tags: [events]
      summary: Get attendance records for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: upsertAttendance
      tags: [events]
      summary: Record or update attendance for a player
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertAttendanceBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── TASKS ────────────────────────────────────────────────────────────────
...
  /teams/{teamId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages for a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Post a message to a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
  /teams/{teamId}/activity:
    get:
      operationId: getTeamActivity
      tags: [dashboard]
      summary: Get recent activity feed for a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Teams ──
...
    Event:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [practice, game, meeting, other]
        location:
          type: ["string", "null"]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        attendingCount:
          type: integer
        totalCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - teamId
        - title
        - type
        - startsAt
        - attendingCount
        - totalCount
        - createdAt
        - updatedAt
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [practice, game, meeting, other]
        location:
          type: ["string", "null"]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - title
        - type
        - startsAt
    UpdateEventBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [practice, game, meeting, other]
        location:
          type: ["string", "null"]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Attendance:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
        status:
          type: string
          enum: [attending, not_attending, maybe, no_response]
        notes:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - eventId
        - playerId
        - playerName
        - status
        - updatedAt
    UpsertAttendanceBody:
      type: object
      properties:
        playerId:
          type: integer
        status:
          type: string
          enum: [attending, not_attending, maybe, no_response]
        notes:
          type: ["string", "null"]
      required:
        - playerId
        - status

    # ── Tasks ──
...
    Message:
      type: object
      properties:
        id:
          type: integer
        teamId:
          type: integer
        senderName:
          type: string
        senderRole:
          type: string
          enum: [coach, player, admin]
        content:
          type: string
        pinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - teamId
        - senderName
        - senderRole
        - content
        - pinned
        - createdAt
    CreateMessageBody:
      type: object
      properties:
        senderName:
          type: string
        senderRole:
          type: string
          enum: [coach, player, admin]
        content:
          type: string
        pinned:
          type: boolean
      required:
        - senderName
        - senderRole
        - content

    # ── Dashboard ──
...
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [event_created, player_added, task_created, task_completed, message_sent, event_updated]
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
        entityId:
          type: ["integer", "null"]
      required:
        - id
        - type
        - description
        - occurredAt